Suggestions Resolved

+++ Frontend/src/app/employees/employees.component.ts
@@ -29,13 +29,16 @@ export class EmployeesComponent implements OnInit {
     localStorage.setItem('editEmployee',employee._id)
     this.router.navigate(['update-employee'])
   }
+
   deleteEmployee(employee:any){
-    console.log(employee);
-    this.employeeService.removeEmployee(employee._id)
-    .subscribe((data) => {
+    if(confirm("Are you sure to delete this employee??")) {
+      this.employeeService.removeEmployee(employee._id)
+      .subscribe((data) => {
       this.ngOnInit()
     })
-
+    }else{
+      this.ngOnInit()
+    }
   }
